requirements.txt
flask
requests
python-dotenv
groq
